Visual changes
Stones were added on top of the first stage.
Torches were added at some points of the first stage.
Danger plates were placed in dangerous places to indicate to the player
that he may die.
Was added to an entrance of the cave at the end of the first stage.
In the end of the first phase the screen darkens gradually.